Fixed Issue #54, Campus nickname alert
Removed the code relating to searching for the campus in the searchFromMenu() function as it was never actually used.
I added that code to the searchRoomNumber() function so that you can now get back to the campus map using the search bar.
Fixed an issue in the getMap() function that was added with the inclusion of the 2000 and 3000 buildings.