Uploaded search/demo automated testing file.
Java class that uses Selenium and JUnit to test search functionality of the GGCMaps web page. The purpose of this class is to demo the tools used for automated testing on this project. This class also tests search functionality by performing a search of one room from each building that supports search functionality.