Fixed safety mode toggle
This commit addresses the ability of the safety mode switch to toggle
the safety icon legend on and off. This was done by modifying the code
in client-js/maps.js. The functionality used to toggle the parking
legend was copied into a new function that is invoked every time the
safety mode toggle is toggled.

Issues to be addressed:

The legend does not contain icons for the safety locations on the maps.
This needs to be fixed in the sass/_nav.scss file using the SVG images
in the repository.